+++ ChangeLog.md
@@ -15,6 +15,8 @@
   * Do not include disabled shield boosters in calculations
   * Add 'Damage dealt' section
   * Add 'Damage received' section
+  * Add 'Piercing' information to hardpoints
+  * Add 'Hardness' information to ship summary
 
 #2.2.5
   * Calculate rate of fire for multi-burst weapons